refactor: replace gorm.Model with application.Model in domain entities
- Replace gorm.Model embedded struct with application.Model in all domain
  entities (product, market, review, coupon, shipping) and persistence models
- Update CLAUDE.md and db.md rules to reference application.Model
- Fix uint(id) → id (int64) in product service to match snowflake.NextID return type

+++ admin/internal/application/product/service.go
@@ -46,7 +46,7 @@ func (s *service) CreateProduct(ctx context.Context, tenantID shared.TenantID, r
 	if err != nil {
 		return nil, err
 	}
-	p.ID = uint(id)
+	p.ID = id
 
 	p.CostPrice = ToDomainMoney(req.CostPrice, req.Currency)
 
@@ -189,7 +189,7 @@ func (s *service) CreateProductWithTx(ctx context.Context, tenantID shared.Tenan
 		if err != nil {
 			return err
 		}
-		p.ID = uint(id)
+		p.ID = id
 
 		p.CostPrice = ToDomainMoney(req.CostPrice, req.Currency)
